Minor fixes

+++ src/flash/display/GraphicsSolidFill.java
@@ -4,7 +4,7 @@ public class GraphicsSolidFill implements IGraphicsData {
 	public int color;
 	public double alpha;
 
-	public GraphicsSolidFill(int i, int i1) {
+	public GraphicsSolidFill(int i, double i1) {
 	}
 
 	public GraphicsSolidFill() {
